src/pyproject.toml
[tool.poetry]
name = "onnxrt-server"
version = "0.1.0"
description = ""
authors = ["Proton <feisuzhu@163.com>"]

[tool.poetry.dependencies]
python = "^3.6"
bjoern = "^3.1.0"
flask = "^1.1.1"
msgpack = "^0.6.2"
numpy = "^1.18.1"
wsgi-request-logger = "^0.4.6"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
